2022 day18 and 19
