--- advent_of_code/aoc/2022/inputs/day18_sample.txt
@@ -1,13 +1,0 @@
-2,2,2
-1,2,2
-3,2,2
-2,1,2
-2,3,2
-2,2,1
-2,2,3
-2,2,4
-2,2,6
-1,2,5
-3,2,5
-2,1,5
-2,3,5
